pyproject.toml
[project]
name = "Sudoku"
version = "0.1.0"
description = "A Sudoku solver and generator"
authors = [
    {name = "Potato"}
]
readme = "README.md"
requires-python = ">=3.9,<4"

dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "authlib (>=1.4.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "starlette-session (>=0.4.3,<0.5.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)",
    "pyngrok (>=7.2.3,<8.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "python-jose (>=3.3.0,<4.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
